requirements.txt
Django==1.6.4
Markdown==2.4.1
Pillow==2.4.0
PyYAML==3.11
South==0.8.4
Unidecode==0.04.14
Whoosh==2.6.0
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.5
django-filter==0.7
django-haystack==2.1.0
django-oauth==1.1
django-phonenumber-field==0.5
django-toolbelt==0.0.1
django-twilio==0.5.1
django-twilio-sms==0.2
djangorestframework==2.3.13
gunicorn==18.0
haystack==0.15
httplib2==0.9
phonenumbers==6.1.0
psycopg2==2.5.2
pystache==0.5.3
python-ptrace==0.7
six==1.6.1
static==1.0.2
twilio==3.6.6
wsgiref==0.1.2
